refactor: move models to domain folder
